test(vm): isolate sdk host wrapper trap
Add focused sdk end-to-end coverage that distinguishes passing noop and panic-abort paths from the still-failing host-wrapper path. Record the current log wrapper ABI investigation in code by aligning the sdk and host binding signatures and keeping the richer reproductions ignored until the underlying unreachable trap is resolved.

+++ blocky-sdk/src/lib.rs
@@ -54,7 +54,7 @@ pub fn decode_args<T: BorshDeserialize>() -> Result<T, String> {
 }
 
 pub fn log(message: &str) {
-    unsafe { env::log(message.as_ptr(), message.len() as i32) }
+    unsafe { env::log(message.as_ptr(), message.len() as i64) }
 }
 
 pub fn balance() -> u64 {
@@ -146,7 +146,7 @@ pub mod env {
         pub fn get_deposit() -> i64;
         #[link_name = "transfer"]
         pub fn env_transfer(to_ptr: *const u8, amount: i64) -> i32;
-        pub fn log(msg_ptr: *const u8, msg_len: i32);
+        pub fn log(msg_ptr: *const u8, msg_len: i64);
         pub fn abort(msg_ptr: *const u8, msg_len: i32) -> !;
     }
 }
@@ -256,7 +256,7 @@ mod tests {
     }
 
     #[unsafe(no_mangle)]
-    extern "C" fn log(msg_ptr: *const u8, msg_len: i32) {
+    extern "C" fn log(msg_ptr: *const u8, msg_len: i64) {
         let bytes = unsafe { std::slice::from_raw_parts(msg_ptr, msg_len.max(0) as usize) };
         host()
             .lock()
